--- src/pages/Review.jsx
@@ -1,6 +1,0 @@
-import ReviewForm from 'components/reviewComponents/ReviewForm';
-import React from 'react';
-
-export const Review = () => {
-  return <ReviewForm />;
-};
