+++ src/pages/ReviewPage.jsx
@@ -1,0 +1,12 @@
+import ReviewForm from 'components/reviewComponents/ReviewForm';
+import { ReviewList } from 'components/reviewComponents/ReviewList';
+import React from 'react';
+
+export const ReviewPage = () => {
+  return (
+    <>
+      <ReviewForm />
+      <ReviewList />
+    </>
+  );
+};
